refactor(handler) merge opentracing.lua and handler.lua
